refactor: drop pinia stores
