+++ ui/nuxt.config.ts
@@ -59,11 +59,12 @@ export default defineNuxtConfig({
     },
     pageTransition: { name: 'page' },
   },
-  modules: ['@nuxt/ui', '@pinia/nuxt', '@vueuse/nuxt', '@nuxt/eslint'],
+  modules: ['@nuxt/ui', '@vueuse/nuxt', '@nuxt/eslint'],
   icon: {
     provider: 'none',
     fallbackToApi: false,
     clientBundle: {
+      icons: isProd ? [] : ['lucide:book'],
       scan: {
         globInclude: ['app/**/*.{vue,ts,js}', 'node_modules/@nuxt/ui/dist/shared/ui*.mjs'],
         globExclude: ['dist', 'build', 'coverage', 'test', 'tests', '.*'],
