+++ ui/package.json
@@ -26,7 +26,6 @@
     "@nuxt/eslint": "^1.15.2",
     "@nuxt/eslint-config": "^1.15.2",
     "@nuxt/ui": "^4.6.1",
-    "@pinia/nuxt": "^0.11.3",
     "@vueuse/core": "^14.2.1",
     "@vueuse/nuxt": "^14.2.1",
     "@xterm/addon-fit": "^0.11.0",
@@ -40,7 +39,6 @@
     "marked-gfm-heading-id": "^4.1.4",
     "moment": "^2.30.1",
     "nuxt": "^4.4.2",
-    "pinia": "^3.0.4",
     "tailwindcss": "^4.2.2",
     "vue": "^3.5.32",
     "vue-router": "^5.0.4"
